split URI parser into perl-lsp-uri microcrate
