+++ crates/perl-lsp-providers/Cargo.toml
@@ -35,6 +35,7 @@ perl-lsp-rename = { workspace = true }
 perl-lsp-navigation = { workspace = true }
 perl-lsp-completion = { workspace = true }
 perl-lsp-code-actions = { workspace = true }
+perl-lsp-uri = { workspace = true }
 rustc-hash = "2.1.1"
 serde_json = "1.0.149"
 lsp-types = { version = "0.97.0", optional = true }
