+++ crates/perl-lsp-providers/src/ide/lsp_compat/uri.rs
@@ -1,20 +1,6 @@
-//! LSP module (deprecated)
+//! LSP URI compatibility module.
 //!
-//! **DEPRECATED**: This module has moved to the `perl-lsp` crate.
-//!
-//! For backwards compatibility during the migration period, this module
-//! is kept as an empty stub. Migrate to `perl_lsp::utils::uri`.
-//!
-//! # Migration
-//!
-//! ```ignore
-//! // Old:
-//! use perl_parser::uri;
-//!
-//! // New:
-//! use perl_lsp::utils::uri;
-//! ```
+//! URI parsing helpers now live in the `perl-lsp-uri` microcrate.
+//! This module re-exports the public API for compatibility.
 
-// This module intentionally has no contents.
-// All functionality has moved to the perl-lsp crate.
-// Direct re-export is not possible due to circular dependency.
+pub use perl_lsp_uri::parse_uri;
