+++ crates/perl-lsp-uri/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "perl-lsp-uri"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Typed URI parsing helpers for perl-lsp"
+readme = "README.md"
+documentation = "https://docs.rs/perl-lsp-uri"
+keywords = ["perl", "lsp", "uri"]
+categories = ["development-tools", "text-editors"]
+include = [
+  "src/**",
+  "Cargo.toml",
+  "README.md",
+  "LICENSE*"
+]
+
+[dependencies]
+lsp-types = "0.97.0"
+
+[lints]
+workspace = true
