+++ crates/perl-lsp-uri/README.md
@@ -1,0 +1,6 @@
+# perl-lsp-uri
+
+Typed URI parsing helpers for the Perl LSP ecosystem.
+
+This microcrate centralizes parsing to `lsp_types::Uri` with a non-panicking
+fallback strategy.
