+++ crates/perl-lsp/Cargo.toml
@@ -28,6 +28,7 @@ perl-lsp-providers = { workspace = true, features = ["lsp-compat"] }
 perl-lsp-formatting = { workspace = true }
 perl-lsp-cancellation = { workspace = true }
 perl-lsp-limits = { workspace = true }
+perl-lsp-uri = { workspace = true }
 perl-lsp-feature-governance = { workspace = true }
 perl-lsp-tooling = { workspace = true }
 perl-lsp-protocol = { workspace = true }
